refactor: repository class package path 변경
