+++ src/main/java/com/zelusik/eatery/app/repository/member/TermsInfoRepository.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.app.repository;
+package com.zelusik.eatery.app.repository.member;
 
 import com.zelusik.eatery.app.domain.TermsInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
