+++ src/main/java/com/zelusik/eatery/app/service/MemberService.java
@@ -11,9 +11,9 @@ import com.zelusik.eatery.app.dto.member.MemberDto;
 import com.zelusik.eatery.app.dto.member.request.MemberUpdateRequest;
 import com.zelusik.eatery.app.dto.member.request.TermsAgreeRequest;
 import com.zelusik.eatery.app.dto.terms_info.TermsInfoDto;
-import com.zelusik.eatery.app.repository.MemberDeletionSurveyRepository;
-import com.zelusik.eatery.app.repository.MemberRepository;
-import com.zelusik.eatery.app.repository.TermsInfoRepository;
+import com.zelusik.eatery.app.repository.member.MemberDeletionSurveyRepository;
+import com.zelusik.eatery.app.repository.member.MemberRepository;
+import com.zelusik.eatery.app.repository.member.TermsInfoRepository;
 import com.zelusik.eatery.global.exception.member.MemberIdNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
