+++ src/main/java/com/zelusik/eatery/app/service/PlaceService.java
@@ -8,8 +8,8 @@ import com.zelusik.eatery.app.dto.place.OpeningHoursTimeDto;
 import com.zelusik.eatery.app.dto.place.PlaceDto;
 import com.zelusik.eatery.app.dto.place.PlaceScrapingInfo;
 import com.zelusik.eatery.app.dto.place.request.PlaceCreateRequest;
-import com.zelusik.eatery.app.repository.OpeningHoursRepository;
-import com.zelusik.eatery.app.repository.PlaceRepository;
+import com.zelusik.eatery.app.repository.place.OpeningHoursRepository;
+import com.zelusik.eatery.app.repository.place.PlaceRepository;
 import com.zelusik.eatery.global.exception.place.PlaceNotFoundException;
 import com.zelusik.eatery.global.exception.scraping.OpeningHoursUnexpectedFormatException;
 import com.zelusik.eatery.global.exception.scraping.ScrapingServerInternalError;
