+++ src/main/java/com/zelusik/eatery/app/service/ProfileImageService.java
@@ -3,7 +3,7 @@ package com.zelusik.eatery.app.service;
 import com.zelusik.eatery.app.domain.member.Member;
 import com.zelusik.eatery.app.domain.member.ProfileImage;
 import com.zelusik.eatery.app.dto.file.S3FileDto;
-import com.zelusik.eatery.app.repository.ProfileImageRepository;
+import com.zelusik.eatery.app.repository.member.ProfileImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
