+++ src/main/java/com/zelusik/eatery/app/service/ReviewFileService.java
@@ -3,7 +3,7 @@ package com.zelusik.eatery.app.service;
 import com.zelusik.eatery.app.domain.Review;
 import com.zelusik.eatery.app.domain.ReviewFile;
 import com.zelusik.eatery.app.dto.file.S3FileDto;
-import com.zelusik.eatery.app.repository.ReviewFileRepository;
+import com.zelusik.eatery.app.repository.review.ReviewFileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
