+++ src/test/java/com/zelusik/eatery/app/service/MemberServiceTest.java
@@ -6,8 +6,8 @@ import com.zelusik.eatery.app.domain.TermsInfo;
 import com.zelusik.eatery.app.dto.member.MemberDto;
 import com.zelusik.eatery.app.dto.member.request.TermsAgreeRequest;
 import com.zelusik.eatery.app.dto.terms_info.TermsInfoDto;
-import com.zelusik.eatery.app.repository.MemberRepository;
-import com.zelusik.eatery.app.repository.TermsInfoRepository;
+import com.zelusik.eatery.app.repository.member.MemberRepository;
+import com.zelusik.eatery.app.repository.member.TermsInfoRepository;
 import com.zelusik.eatery.global.exception.member.MemberIdNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
