+++ src/test/java/com/zelusik/eatery/app/service/ReviewFileServiceTest.java
@@ -1,7 +1,7 @@
 package com.zelusik.eatery.app.service;
 
 import com.zelusik.eatery.app.domain.Review;
-import com.zelusik.eatery.app.repository.ReviewFileRepository;
+import com.zelusik.eatery.app.repository.review.ReviewFileRepository;
 import com.zelusik.eatery.util.MultipartFileTestUtils;
 import com.zelusik.eatery.util.ReviewTestUtils;
 import com.zelusik.eatery.util.S3FileTestUtils;
